finishing up

+++ Midnight/Core/data.h
@@ -12,7 +12,7 @@
 namespace aveng {
 	const int RIGHT = -1;
 	const int LEFT = 1;
-	const int PI = 3.14159265f; // 3589793238462643383279502884197969399375105820974944592 check yo'self
+	const float PI = 3.14159265f; // 3589793238462643383279502884197969399375105820974944592 check yo'self
 	const float viewRadius{ .5f };	// Radius of the invisible sphere for which our viewer is at the origin
 
 	//std::unordered_map < std::string , AvengModel > models_in_scene;
